Refactored pytest fixtures.
